moved data sidecar verification to fork_choice_store
