+++ fork_choice_store/Cargo.toml
@@ -13,6 +13,8 @@ clock = { workspace = true }
 crossbeam-skiplist = { workspace = true }
 derivative = { workspace = true }
 derive_more = { workspace = true }
+educe = { workspace = true }
+eip_7594 = { workspace = true }
 eth2_libp2p = { workspace = true }
 execution_engine = { workspace = true }
 features = { workspace = true }
