+++ operation_pools/Cargo.toml
@@ -13,7 +13,6 @@ bls = { workspace = true }
 clock = { workspace = true }
 conv = { workspace = true }
 dedicated_executor = { workspace = true }
-eip_7594 = { workspace = true }
 eth1_api = { workspace = true }
 eth2_libp2p = { workspace = true }
 features = { workspace = true }
